🚀 Fix: Resolución de errores de tipos (int/double) y sincronización del protocolo 'disfrute' para Web

+++ apps/finance_copilot/lib/presentation/dashboard/dashboard_screen.dart
@@ -63,7 +63,7 @@ class _DashboardScreenState extends ConsumerState<DashboardScreen> {
             children: [
               _buildSidebar(),
               Expanded(
-                child: _buildMainContent(context, peacePoint, sustainability, protocol, totalIncome, totalExpenses),
+                child: _buildMainContent(context, peacePoint.toDouble(), sustainability, protocol, totalIncome, totalExpenses),
               ),
             ],
           ),
